+++ src/main/java/ru/yandex/practicum/filmorate/service/impl/ReviewLikesServiceImpl.java
@@ -11,19 +11,19 @@ public class ReviewLikesServiceImpl implements ReviewLikesService {
 
     private final ReviewLikesStorage reviewLikesStorage;
 
-    public void like(int reviewId, int userId) {
+    public void like(Long reviewId, Long userId) {
         reviewLikesStorage.like(reviewId, userId);
     }
 
-    public void dislike(int reviewId, int userId) {
+    public void dislike(Long reviewId, Long userId) {
         reviewLikesStorage.dislike(reviewId, userId);
     }
 
-    public void deleteLike(int reviewId, int userId) {
+    public void deleteLike(Long reviewId, Long userId) {
         reviewLikesStorage.deleteLike(reviewId, userId);
     }
 
-    public void deleteDislike(int reviewId, int userId) {
+    public void deleteDislike(Long reviewId, Long userId) {
         reviewLikesStorage.deleteDislike(reviewId, userId);
     }
 }
