+++ src/main/java/ru/yandex/practicum/filmorate/storage/impl/ReviewLikesStorageImpl.java
@@ -12,7 +12,7 @@ public class ReviewLikesStorageImpl implements ReviewLikesStorage {
     private final JdbcTemplate jdbcTemplate;
 
     @Override
-    public void like(int reviewId, int userId) {
+    public void like(Long reviewId, Long userId) {
         String sql =
                 "INSERT INTO review_likes (review_id, is_like, user_id) " +
                         "VALUES (?, ?, ?)";
@@ -22,7 +22,7 @@ public class ReviewLikesStorageImpl implements ReviewLikesStorage {
     }
 
     @Override
-    public void dislike(int reviewId, int userId) {
+    public void dislike(Long reviewId, Long userId) {
         String sql =
                 "INSERT INTO review_likes (review_id, is_like, user_id) " +
                         "VALUES (?, ?, ?)";
@@ -32,7 +32,7 @@ public class ReviewLikesStorageImpl implements ReviewLikesStorage {
     }
 
     @Override
-    public void deleteLike(int reviewId, int userId) {
+    public void deleteLike(Long reviewId, Long userId) {
         String sql =
                 "DELETE FROM review_likes WHERE review_id = ? AND is_like = ? AND user_id = ?";
 
@@ -41,7 +41,7 @@ public class ReviewLikesStorageImpl implements ReviewLikesStorage {
     }
 
     @Override
-    public void deleteDislike(int reviewId, int userId) {
+    public void deleteDislike(Long reviewId, Long userId) {
         String sql =
                 "DELETE FROM review_likes WHERE review_id = ? AND is_like = ? AND user_id = ?";
 
@@ -49,7 +49,7 @@ public class ReviewLikesStorageImpl implements ReviewLikesStorage {
         updateReviewUsefulness(reviewId, 1);
     }
 
-    private void updateReviewUsefulness(int reviewId, int value) {
+    private void updateReviewUsefulness(Long reviewId, int value) {
         String sql = "UPDATE reviews SET useful = useful + ? WHERE id = ?";
         jdbcTemplate.update(sql, value, reviewId);
     }
